refactor: Improved separation of concerns in NavHost.kt
- Added currency indicator to composables in ItemDetailsCard.kt, InventoryItem.kt, and ItemInputForm.kt for improved uniformity.
- Moved `ItemDetailsScreen()` from `InventoryNavHost` to a separate `ItemDetailsScreenCollector` composable to clarify the structure of `NavHost.kt` and improve proper separation of concerns.
